rpm_rbcpr_stats: Fix the end address for the resource
Currently the driver assumes a fixed end address for the resource
and validates the rbcpr stats start address against this end address.
The RBCPR start address is now dynamically read from the image header
and the end address can no longer be assumed to be fixed.

Update the resource end address to be at 2K offset from the start
address before ioremap.

Change-Id: I6641997260582242f2211d6801b701c5550cecd4
Signed-off-by: Archana Sathyakumar <asathyak@codeaurora.org>